+++ oauth2/src/test/scala/org/broadinstitute/dsde/workbench/oauth2/OpenIDConnectAkkaHttpSpec.scala
@@ -56,6 +56,24 @@ class OpenIDConnectAkkaHttpSpec extends AnyFlatSpecLike with Matchers with Workb
     res.unsafeRunSync()
   }
 
+  it should "redirect to a different policy" in {
+    val res = for {
+      config <- OpenIDConnectConfiguration[IO](
+        "https://terradevb2c.b2clogin.com/terradevb2c.onmicrosoft.com/v2.0?p=b2c_1a_signup_signin",
+        ClientId("some_client")
+      )
+      req = Get(Uri("/oauth2/authorize").withQuery(Query("""p=some-other-policy""")))
+      _ <- req ~> config.oauth2Routes ~> checkIO {
+        handled shouldBe true
+        status shouldBe StatusCodes.Found
+        header[Location].map(_.value) shouldBe Some(
+          "https://terradevb2c.b2clogin.com/terradevb2c.onmicrosoft.com/oauth2/v2.0/authorize?p=some-other-policy"
+        )
+      }
+    } yield ()
+    res.unsafeRunSync()
+  }
+
   it should "reject non-GET requests" in {
     val res = for {
       config <- OpenIDConnectConfiguration[IO]("https://accounts.google.com", ClientId("some_client"))
@@ -107,6 +125,57 @@ class OpenIDConnectAkkaHttpSpec extends AnyFlatSpecLike with Matchers with Workb
         status shouldBe StatusCodes.OK
         val resp = responseAs[String]
         resp shouldBe (Map("token" -> "a-token")).asJson.noSpaces
+      }
+    } yield ()
+
+    mockServer.use(_ => res).unsafeRunSync()
+  }
+
+  it should "proxy requests with a policy field" in {
+    val formData = Map("grant_type" -> "authorization_code",
+                       "code" -> "1234",
+                       "client_id" -> "some_client",
+                       "p" -> "some-other-policy"
+    )
+    val backendRoute = path(".well-known" / "openid-configuration") {
+      get {
+        complete {
+          Map("issuer" -> "unused",
+              "authorization_endpoint" -> "unused",
+              "token_endpoint" -> "http://localhost:9000/token"
+          )
+        }
+      }
+    } ~
+      path("token") {
+        post {
+          formFieldSeq { fields =>
+            parameterSeq { params =>
+              fields.toMap shouldBe formData
+              params.toMap shouldBe Map("p" -> "some-other-policy")
+              complete(Map("token" -> "a-token"))
+            }
+          }
+        }
+      }
+    val mockServer =
+      Resource.make(IO.fromFuture(IO(Http().newServerAt("0.0.0.0", 9000).bind(backendRoute))))(serverBinding =>
+        IO.fromFuture(IO(serverBinding.terminate(3.seconds))).void
+      )
+
+    val res = for {
+      config <- OpenIDConnectConfiguration[IO]("http://localhost:9000",
+                                               ClientId("some_client"),
+                                               Some(ClientSecret("some_long_client_secret"))
+      )
+      req = Post("/oauth2/token").withEntity(
+        FormData(formData).toEntity
+      )
+      _ <- req ~> config.oauth2Routes ~> checkIO {
+        handled shouldBe true
+        status shouldBe StatusCodes.OK
+        val resp = responseAs[String]
+        resp shouldBe Map("token" -> "a-token").asJson.noSpaces
       }
     } yield ()