+++ oauth2/CHANGELOG.md
@@ -7,12 +7,15 @@ This file documents changes to the `workbench-oauth2` library, including notes o
 Changed:
 - Changed method on `OpenIDConnectConfiguration` to return `OpenIDProviderMetadata` case class instead of strings
 - Fixed issue causing akka-http `Stream cannot be materialized more than once` on `/oauth2/token` endpoint
+- Support authority endpoints with a dynamic policy on the query string, e.g.:
+   - https://terradevb2c.b2clogin.com/terradevb2c.onmicrosoft.com/v2.0?p=B2C_1A_SIGNUP_SIGNIN
+- Update swagger-ui to 4.11.1
 
 Added:
 - Added methods for clientId and authorityEndpoint to `OpenIDConnectConfiguration`
 - Added akka-http route `/oauth2/configuration` which returns JSON containing the clientId and authorityEndpoint
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.2-1962b9a"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.2-TRAVIS-REPLACE-ME"`
 
 ## 0.1